pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "tender-erp"
version = "0.5.0"
description = "Tender & Compliance Management Desktop Application"
requires-python = ">=3.11"
authors = [{name = "Aadi Pandey"}]
dependencies = [
    "PySide6>=6.6.0",
    "SQLAlchemy>=2.0.0",
    "argon2-cffi>=23.1.0",
    "cryptography>=42.0.0",
    "openpyxl>=3.1.0",
    "reportlab>=4.0.0",
    "plyer>=2.1.0",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = ["pytest>=7.4", "pytest-qt>=4.2"]

[project.scripts]
tender-erp = "tender_erp.__main__:main"
tender-erp-cli = "tender_erp.cli:main"

[tool.setuptools.packages.find]
include = ["tender_erp*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
